+++ ts/packages/knowProTest/src/knowproContext.ts
@@ -3,15 +3,6 @@
 
 import { ChatModel, openai } from "aiclient";
 import * as kp from "knowpro";
-import * as cm from "conversation-memory";
-import {
-    AnswerDebugContext,
-    GetAnswerRequest,
-    GetAnswerResponse,
-    SearchRequest,
-    SearchResponse,
-} from "./types.js";
-import { error, Result, success } from "typechat";
 
 export class KnowproContext {
     public knowledgeModel: ChatModel;
@@ -37,176 +28,10 @@ export class KnowproContext {
         }
         return this.conversation!;
     }
-
-    public async execSearchRequest(
-        request: SearchRequest,
-    ): Promise<SearchResponse> {
-        const conversation = this.ensureConversationLoaded();
-        const searchText = request.query;
-        const debugContext: AnswerDebugContext = { searchText };
-
-        const options: kp.LanguageSearchOptions = {
-            ...createSearchOptions(request),
-            compileOptions: {
-                exactScope: request.exactScope,
-                applyScope: request.applyScope,
-            },
-        };
-        options.exactMatch = request.exact;
-        if (request.fallback) {
-            options.fallbackRagOptions = {
-                maxMessageMatches: options.maxMessageMatches,
-                maxCharsInBudget: options.maxCharsInBudget,
-                thresholdScore: 0.7,
-            };
-        }
-        const langFilter = createLangFilter(undefined, request);
-        const searchResults =
-            conversation instanceof cm.Memory
-                ? await conversation.searchWithLanguage(
-                      searchText,
-                      options,
-                      langFilter,
-                      debugContext,
-                  )
-                : await kp.searchConversationWithLanguage(
-                      conversation,
-                      searchText,
-                      this.queryTranslator,
-                      options,
-                      langFilter,
-                      debugContext,
-                  );
-
-        return { searchResults, debugContext };
-    }
-
-    public async execGetAnswerRequest(
-        request: GetAnswerRequest,
-        progressCallback?: (
-            index: number,
-            question: string,
-            answer: Result<kp.AnswerResponse>,
-        ) => void,
-    ): Promise<GetAnswerResponse> {
-        const searchResponse = await this.execSearchRequest(request);
-        const searchResults = searchResponse.searchResults;
-        const response: GetAnswerResponse = {
-            searchResponse,
-            answerResponses: error("Not initialized"),
-        };
-        if (!searchResults.success) {
-            return response;
-        }
-        if (!kp.hasConversationResults(searchResults.data)) {
-            return response;
-        }
-        const answerResponses = await this.getAnswersForSearchResults(
-            request,
-            searchResults.data,
-            progressCallback,
-        );
-        response.answerResponses = answerResponses;
-        return response;
-    }
-
-    public async getAnswersForSearchResults(
-        request: GetAnswerRequest,
-        searchResults: kp.ConversationSearchResult[],
-        progressCallback?: (
-            index: number,
-            question: string,
-            answer: Result<kp.AnswerResponse>,
-        ) => void,
-    ): Promise<Result<kp.AnswerResponse[]>> {
-        const conversation = this.ensureConversationLoaded();
-        const fastStopSav = this.answerGenerator.settings.fastStop;
-        if (request.fastStop) {
-            this.answerGenerator.settings.fastStop = request.fastStop;
-        }
-        let answerResponses: kp.AnswerResponse[] = [];
-        try {
-            if (!request.messages) {
-                // Don't include raw message text... try answering only with knowledge
-                searchResults.forEach((r) => (r.messageMatches = []));
-            }
-            const choices = request.choices?.split(";");
-            const options = createAnswerOptions(request);
-            for (let i = 0; i < searchResults.length; ++i) {
-                const searchResult = searchResults[i];
-                let question = searchResult.rawSearchQuery ?? request.query;
-                if (choices && choices.length > 0) {
-                    question = kp.createMultipleChoiceQuestion(
-                        question,
-                        choices,
-                    );
-                }
-                const answerResult = await kp.generateAnswer(
-                    conversation,
-                    this.answerGenerator,
-                    question,
-                    searchResult,
-                    undefined,
-                    options,
-                );
-                if (!answerResult.success) {
-                    return answerResult;
-                }
-                answerResponses.push(answerResult.data);
-                if (progressCallback) {
-                    progressCallback(i, question, answerResult);
-                }
-            }
-        } finally {
-            this.answerGenerator.settings.fastStop = fastStopSav;
-        }
-        return success(answerResponses);
-    }
 }
 
 export function createKnowledgeModel() {
     const chatModelSettings = openai.apiSettingsFromEnv(openai.ModelType.Chat);
     chatModelSettings.retryPauseMs = 10000;
     return openai.createJsonChatModel(chatModelSettings, ["knowproTest"]);
-}
-
-function createLangFilter(
-    when: kp.WhenFilter | undefined,
-    request: SearchRequest,
-): kp.LanguageSearchFilter | undefined {
-    if (request.ktype) {
-        when ??= {};
-        when.knowledgeType = request.ktype;
-    }
-    if (request.tag) {
-        when ??= {};
-        when.tags = [request.tag];
-    }
-    if (request.thread) {
-        when ??= {};
-        when.threadDescription = request.thread;
-    }
-    return when;
-}
-
-function createSearchOptions(request: SearchRequest): kp.SearchOptions {
-    let options = kp.createSearchOptions();
-    options.exactMatch = request.exact;
-    options.maxMessageMatches = request.messageTopK;
-    options.maxCharsInBudget = request.charBudget;
-    return options;
-}
-
-function createAnswerOptions(
-    namedArgs: GetAnswerRequest,
-): kp.AnswerContextOptions {
-    let topK = namedArgs.knowledgeTopK;
-    if (topK === undefined) {
-        return {};
-    }
-    const options: kp.AnswerContextOptions = {
-        entitiesTopK: topK,
-        topicsTopK: topK,
-    };
-    return options;
 }
